py3DN v1.04
Py3DN, is an open-source solution providing innovative tools to analyze
3D data collected with the widely used commercial system Neurolucida
(MBF). It allows the construction of mathematical representations of
neuronal topology, visualization and a variety of morphometric analysis
on the neuronal structures. Above all, it provides a flexible
environment where new types of analyses can be easily set up allowing
total freedom to test new hypotheses. The application was developed in
Python and uses Blender (an open source program) to produce detailed 3D
representations of raw and processed data.
Py3DN is described in the paper:
Aguiar P et al, "Versatile morphometric analysis and visualization of
the three-dimensional structure of neurons", Neuroinformatics, 2013,
11(4):393-403
doi: 10.1007/s12021-013-9188-z